scripts/requirements.txt
deep-translator
PyYAML